requirements/base.txt
pip-chill==1.0.3
psycopg2-binary==2.9.9
python-dotenv==1.0.0
django-cors-headers==4.3.0
django-filter==23.3
djangorestframework-simplejwt==5.3.0
drf-spectacular==0.26.5